move all strategy-related types to Strategies subfolder and namespace
