+++ src/libraries/System.Private.CoreLib/src/System/IO/FileStream.cs
@@ -1,6 +1,7 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
+using System.IO.Strategies;
 using System.Runtime.Serialization;
 using System.Runtime.Versioning;
 using System.Threading;
