+++ src/libraries/System.Private.CoreLib/src/System/IO/Strategies/LegacyFileStreamStrategy.Lock.Unix.cs
@@ -1,7 +1,7 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-namespace System.IO
+namespace System.IO.Strategies
 {
     internal sealed partial class LegacyFileStreamStrategy : FileStreamStrategy
     {
